Adds tests for the requiredblocks element
